+++ Solldaten.h
@@ -16,7 +16,7 @@
 #include <cassert>
 
 
-pcl::PointCloud<pcl::PointXYZ>::Ptr sollstein_erstellen(float aufloesung = 0.1, std::string dat="leer");
+pcl::PointCloud<pcl::PointXYZ>::Ptr sollstein_erstellen(double aufloesung = 0.1, std::string dat="leer");
 std::vector<float> ausgabe_solldaten(std::string dat);
 
 typedef struct points_txt_soll
